Applied fix to rigid_wasabi as well

+++ src/models/rigid_wasabi_model.py
@@ -97,7 +97,7 @@ class rigid_wasabi_model(pipeline_pytorch_model):
             with torch.no_grad():
                 for i in range(0, len(norm_data), batch_size):
                     prediction = self.model.model(norm_patches[i:i+batch_size], legend_patches[i:i+batch_size])
-                    prediction = torch.argmax(prediction, dim=1).cpu().numpy().astype(np.uint8)
+                    prediction = torch.argmax(prediction, dim=1).cpu().numpy()
                     
                     prediction_patches.append(prediction)
                     
@@ -132,7 +132,7 @@ class rigid_wasabi_model(pipeline_pytorch_model):
             cur_max = cur_max
 
             for k in predictions:
-                predictions[k] = (predictions[k] >= cur_max) & (cur_max > 0.5)
+                predictions[k] = ((predictions[k] >= cur_max) & (cur_max > 0.5)).astype(np.uint8)
 
         return predictions
